requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
jinja2==3.1.2
tinkoff-investments==0.2.0b67
python-dotenv==1.0.0
pydantic==1.10.12
APScheduler==3.10.4
pytz==2023.3
